+++ packages/liveblocks-core/src/crdts/__tests__/LiveObject.mockserver.test.ts
@@ -144,7 +144,7 @@ describe("LiveObject edge cases", () => {
         modified: {
           node: obj,
           type: "LiveObject",
-          updates: { b: { type: "delete" } },
+          updates: { b: { type: "delete", deletedItem: secondItem } },
         },
         reverse: [
           {